refactor(cockpit): move composition tests to owner module
